fix: Restrict seed file to handle test and dev environments

+++ priv/repo/seeds.exs
@@ -1,7 +1,8 @@
 alias Realtime.{Api.Tenant, Repo}
 import Ecto.Adapters.SQL, only: [query: 3]
 
-tenant_name = "realtime-dev-tenant"
+tenant_name = "realtime-dev"
+default_db_host = if Mix.env() in [:dev, :test], do: "localhost", else: "host.docker.internal"
 
 Repo.transaction(fn ->
   case Repo.get_by(Tenant, external_id: tenant_name) do
@@ -13,16 +14,17 @@ Repo.transaction(fn ->
   |> Tenant.changeset(%{
     "name" => tenant_name,
     "external_id" => tenant_name,
-    "jwt_secret" => "super-secret-jwt-token-with-at-least-32-characters-long",
+    "jwt_secret" =>
+      System.get_env("API_JWT_SECRET", "super-secret-jwt-token-with-at-least-32-characters-long"),
     "extensions" => [
       %{
         "type" => "postgres_cdc_rls",
         "settings" => %{
-          "db_name" => "postgres",
-          "db_host" => "localhost",
-          "db_user" => "postgres",
-          "db_password" => "postgres",
-          "db_port" => "5432",
+          "db_name" => System.get_env("DB_NAME", "postgres"),
+          "db_host" => System.get_env("DB_HOST", default_db_host),
+          "db_user" => System.get_env("DB_USER", "postgres"),
+          "db_password" => System.get_env("DB_PASSWORD", "postgres"),
+          "db_port" => System.get_env("DB_PORT", "5432"),
           "region" => "us-east-1",
           "poll_interval_ms" => 100,
           "poll_max_record_bytes" => 1_048_576,
@@ -34,21 +36,20 @@ Repo.transaction(fn ->
   |> Repo.insert!()
 end)
 
-publication = "supabase_realtime"
+if Mix.env() in [:dev, :test] do
+  publication = "supabase_realtime"
 
-{:ok, _} =
-  Repo.transaction(fn ->
-    [
-      "drop publication if exists #{publication}",
-      "drop table if exists public.test_tenant;",
-      "create table public.test_tenant (
-        id SERIAL PRIMARY KEY,
-        details text
-        );",
-      "grant all on table public.test_tenant to anon;",
-      "grant all on table public.test_tenant to postgres;",
-      "grant all on table public.test_tenant to authenticated;",
-      "create publication #{publication} for table public.test_tenant"
-    ]
-    |> Enum.each(&query(Repo, &1, []))
-  end)
+  {:ok, _} =
+    Repo.transaction(fn ->
+      [
+        "drop publication if exists #{publication}",
+        "drop table if exists public.test_tenant;",
+        "create table public.test_tenant ( id SERIAL PRIMARY KEY, details text );",
+        "grant all on table public.test_tenant to anon;",
+        "grant all on table public.test_tenant to postgres;",
+        "grant all on table public.test_tenant to authenticated;",
+        "create publication #{publication} for table public.test_tenant"
+      ]
+      |> Enum.each(&query(Repo, &1, []))
+    end)
+end
